// enums/IRMNum.ts
enum IRMInss {
  FIRST_PERCENT = 0.075,
  SECOND_PERCENT = 0.09,
  THIRD_PERCENT = 0.12,
  FOURTY_PERCENT = 0.14,
}

enum IRMDeducao {
  FIRST_DEDUCTION = 142.8,
  SECOND_DEDUCTION = 354.8,
  THIRD_DEDUCTION = 636.13,
  FOURTY_DEDUCTION = 869.36,
}

enum IRMDepen {
  dependentes = 189.59,
}

enum IRMPorcent {
  ZERO_RANGE = 0,
  FIRST_RANGE = 0.075,
  SECOND_RANGE = 0.15,
  THIRD_RANGE = 0.225,
  FOURTY_RANGE = 0.275,
}

export { IRMInss, IRMDeducao, IRMDepen, IRMPorcent };